+++ StorageAccounting.DAL/Models/Storage/Arrival.cs
@@ -8,11 +8,11 @@ namespace StorageAccounting.Domain.Models.Storage;
 /// </summary>
 public class Arrival
 {
-    public double Id { get; set; }
+    public long Id { get; set; }
     
-    public double DocumentId { get; set; }
+    public long DocumentId { get; set; }
 
-    public int PartnerId { get; set; }
+    public long PartnerId { get; set; }
 
     public int PlaceId { get; set; }
